Add files via upload
This is a special kind of conky which gives you complete stats of your system, it was a major development done by Purnomo Hadi which had outdated conky syntax. I renewed the conky syntax and added lot more features to it and gave it a new name. This is the beauty of opensource. Enjoy! ;)