~CLICK HERE TO READ THE RELEASE INFO~
This is a special kind of conky which gives you complete stats of your system, it was a major development done by Purnomo Hadi which had outdated Conky syntax. I renewed the Conky syntax and added lot more features to it and gave it a new name. This is the beauty of open-source. Make sure you have Roboto Mono fonts installed else you can change the fonts according to what fonts you have. Also the GPU part is coded according to a single core GPU, if you have a multi-core GPU you can append the extra core according to your requirements. Enjoy! ;)