CLICK HERE TO READ MAJOR RELEASE INFO!
This Conky configuration was half developed by Purnomo Hadi and that also was written in old Conky syntax that supported in Conky<1.10, I renewed the syntax and ported it for conky>=1.10 and then added major features to it and completed this release. If you don't have Roboto Mono fonts installed then you are free to use any fonts. You are also free to redistribute this unless you give due credits to its authors. Also the GPU part is coded according to a single core GPU, if you have higher GPUs with multiple cores then you can add an extra line to make it supportive. Enjoy!
 master